rag/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
opensearch-py
requests # Necessário para fazer chamadas HTTP diretas ao Ollama